Disable "Convert to Primary Constructor" warning (#216)
Me and other .NET developers hold the stance that primary constructors
are not ready for production use, particularly with dependency
injection. There are debates regarding the styling of primary
constructors, but the bigger issue is that a primary constructor
parameter cannot be made `readonly` (which is crucial with dependency
injection). The inspection "Convert to Primary Constructor" was disabled
in ReSharper CLI when we updated to .NET 8, but the warning is still
present for ReSharper and Rider users. This PR disables this warning to
avoid developers accidentally using a primary constructor.